+++ assets/javascripts/kitten/karl/pages/manifesto/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Hero from './components/hero'
+
+const Manifesto = () => (
+  <>
+    <Hero />
+  </>
+)
+
+export default Manifesto